# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/recepie/{recepie_id}/ingredient": {
            "get": {
                "description": "get all Recepie Ingredients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie/Ingredient"
                ],
                "summary": "Get all Ingredients",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "recepie_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ingredient"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recepie/{recepie_id}/ingredient/": {
            "post": {
                "description": "create Recepie Ingredient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie/Ingredient"
                ],
                "summary": "Create Ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "recepie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createIngredient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recepie/{recepie_id}/ingredient/{ingredient_id}": {
            "get": {
                "description": "get Recepie Ingredient by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie/Ingredient"
                ],
                "summary": "Get Ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "recepie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "ingredient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "update Recepie Ingredient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie/Ingredient"
                ],
                "summary": "Update Ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "recepie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "ingredient_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateIngredient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete Recepie Ingredient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie/Ingredient"
                ],
                "summary": "Delete Ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "recepie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "ingredient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createIngredient": {
            "type": "object",
            "required": [
                "amount",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "amount_type": {
                    "$ref": "#/definitions/models.IngredientAmountType"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handlers.updateIngredient": {
            "type": "object",
            "required": [
                "amount",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "amount_type": {
                    "$ref": "#/definitions/models.IngredientAmountType"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Ingredient": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "amount_type": {
                    "$ref": "#/definitions/models.IngredientAmountType"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.IngredientAmountType": {
            "type": "string",
            "enum": [
                "amount",
                "gramms"
            ],
            "x-enum-varnames": [
                "PIECES",
                "GRAMMS"
            ]
        },
...
    },
    "securityDefinitions": {
        "CookieSID": {
            "type": "apiKey",
            "name": "sid",
            "in": "cookie"
        }
    }
}